fix: sixteenth review round — grid-relative Prony fallback, consumption-matched certificates, D11 decoupling
All four findings reproduced before changing anything, then adopted.

## Prony fallback seeds are grid-relative (P2)

The fallback seeded (beta, omega) = (1, 1) in absolute rate units, so on a
valid non-uniform grid spanning t = 1e7 the M3b fit started seven orders of
magnitude off and least-squares "converged" (success=True) onto the seed
itself -- reproduced: [A, beta, omega] ~ [0.006, 1, 1] for true values 5e-8
and 2e-6, corrupting the fitted rate and suppressing M3b/A8 purely because
of the rate unit. Fallback rates are now 5.0/t_span and the success-path
positivity floors 5e-6/t_span, both reproducing the historical values
exactly at the canonical t_span = 5 (same convention as
ALPHA_SEED_FLOOR_FRAC / M3A_SLOPE_SEED_FRAC). The measured non-uniform
scenario is now unit-invariant across c in [1e-6, 1e6] (pinned end to end).

## D24 uses the eigenvalue certificate when only the gap is missing (P2)

The recomputation required certified_eig even when the caller supplied
petermann_factor -- but that path consumes no eigenvectors. Reproduced: a
stiff network whose eigenvalue certificate resolves a usable gap of
3.32e-6 while only the round-15 eigenvector gate fails returned an
unconverged NaN record. The certificate now matches what is consumed:
certified_eigvals for the gap-only path, certified_eig when the Petermann
factor is recomputed (that path stays fail-closed, pinned).

## D11 decoupled from the D9 vector gate (P2)

The D9 NaN sentinel flowed into bohr_arithmetic_progression, which
silently filtered it and reported the default length 1 as a measured
value. Reproduced: a ladder-Hamiltonian network whose certified spectrum
carries a length-3 progression while D9 is withheld reported 1. D11's
input is now recomputed from the certified spectrum (D11 consumes only
eigenvalues); when the eigenvalues themselves are unresolved,
bohr_ap_length is NaN (field now float-typed, NaN = documented sentinel).

## Release-note correction (P2, documentation)

The atol opt-in for the legacy absolute Hermiticity gate exists on the
standalone is_hermitian predicate only; the builders never accepted it.
The #109 CHANGELOG entry now says so explicitly instead of advertising an
unavailable builder argument.

Full suite 942 passed (936 + 6 new pins); anchors byte-identical; ruff and
mypy clean; CHANGELOG and CITATION.cff pending block updated.

+++ src/liouscope/_types.py
@@ -144,7 +144,7 @@ class NonNormalityResult:
     petermann_max: float        # D9 max K_j
     petermann_factors: np.ndarray
     kreiss: float               # D10 (legacy absolute-grid lower bound)
-    bohr_ap_length: int         # D11 Bohr arithmetic-progression depth
+    bohr_ap_length: float       # D11 Bohr AP depth (int-valued; NaN = withheld, round-16)
     bohr_ap_pauli_bound: float
     henrici_relative: float = float("nan")   # D8b eta_N / ||L||_F in [0, 1]
     kreiss_scaled: float = float("nan")      # D10b scale-relative grid lower bound
